+++ Sources/BraintreeCore/Public/BraintreeCore/BTAppSwitch.h
@@ -8,7 +8,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  Handles return URLs when returning from app switch and routes the return URL to the correct app switch handler class.
- @note `returnURLScheme` must contain your app's registered URL Type that starts with the app's bundleID. When your app returns from app switch, the app delegate should call `handleOpenURL:sourceApplication:`
+ @note `returnURLScheme` must contain your app's registered URL Type that starts with the app's bundleID. When your app returns from app switch, the app delegate should call `handleOpenURL:` or `handleOpenURLContext:`
 */
 @interface BTAppSwitch : NSObject
 
@@ -38,19 +38,9 @@ NS_ASSUME_NONNULL_BEGIN
  Handles a return from app switch
 
  @param url The URL that was opened to return to your app
- @param sourceApplication The source app that requested the launch of your app
  @return `YES` if the app switch successfully handled the URL, or `NO` if the attempt to handle the URL failed.
 */
-+ (BOOL)handleOpenURL:(NSURL *)url sourceApplication:(nullable NSString *)sourceApplication;
-
-/**
- Handles a return from app switch
-
- @param url The URL that was opened to return to your app
- @param options The options dictionary provided by `application:openURL:options:`
- @return `YES` if the app switch successfully handled the URL, or `NO` if the attempt to handle the URL failed.
-*/
-+ (BOOL)handleOpenURL:(NSURL *)url options:(NSDictionary *)options;
++ (BOOL)handleOpenURL:(NSURL *)url;
 
 /**
  Handles a return from app switch
@@ -193,7 +183,7 @@ typedef NS_ENUM(NSInteger, BTAppSwitchTarget) {
 
 /**
  A protocol for handling the return from switching out of an app to gather payment information.
- @note The app may switch out to Mobile Safari or to a native app.
+ @note The app may switch out to SFSafariViewController or to a native app.
 */
 @protocol BTAppSwitchHandler
 
@@ -202,16 +192,15 @@ typedef NS_ENUM(NSInteger, BTAppSwitchTarget) {
 /**
  Determine whether the app switch return URL can be handled.
 
- @param url the URL you receive in `application:openURL:sourceApplication:annotation` when returning to your app
- @param sourceApplication The source application you receive in `application:openURL:sourceApplication:annotation`
+ @param url the URL you receive in  `scene:openURLContexts:` (or `application:openURL:options:` if iOS 12) when returning to your app
  @return `YES` when the object can handle returning from the application with a URL
 */
-+ (BOOL)canHandleAppSwitchReturnURL:(NSURL *)url sourceApplication:(NSString * _Nullable)sourceApplication;
++ (BOOL)canHandleAppSwitchReturnURL:(NSURL *)url;
 
 /**
- Pass control back to `BTPayPalDriver` after returning from app or browser switch.
+ Complete payment flow after returning from app or browser switch.
 
- @param url The URL you receive in `application:openURL:sourceApplication:annotation`
+ @param url The URL you receive in `scene:openURLContexts:` (or `application:openURL:options:` if iOS 12)
 */
 + (void)handleAppSwitchReturnURL:(NSURL *)url;