+++ Sources/BraintreeVenmo/BTVenmoAppSwitchReturnURL.m
@@ -10,10 +10,8 @@ NSString *const BTVenmoAppSwitchReturnURLErrorDomain = @"com.braintreepayments.B
 
 @implementation BTVenmoAppSwitchReturnURL
 
-+ (BOOL)isValidURL:(NSURL *)url sourceApplication:(NSString *)sourceApplication {
-    return [self isVenmoSourceApplication:sourceApplication] || // Pre iOS 13
-           [self isValidVenmoURL:url andUnknownApplication:sourceApplication] || // iOS 13+
-           [self isFakeWalletURL:url andSourceApplication:sourceApplication];
++ (BOOL)isValidURL:(NSURL *)url {
+    return [url.host isEqualToString:@"x-callback-url"] && [url.path hasPrefix:@"/vzero/auth/venmo/"];
 }
 
 - (instancetype)initWithURL:(NSURL *)url {
@@ -36,20 +34,6 @@ NSString *const BTVenmoAppSwitchReturnURLErrorDomain = @"com.braintreepayments.B
         }
     }
     return self;
-}
-
-#pragma mark Internal Helpers
-
-+ (BOOL)isVenmoSourceApplication:(NSString *)sourceApplication {
-    return [sourceApplication hasPrefix:@"net.kortina.labs.Venmo"];
-}
-
-+ (BOOL)isFakeWalletURL:(NSURL *)url andSourceApplication:(NSString *)sourceApplication {
-    return [sourceApplication isEqualToString:@"com.paypal.PPClient.Debug"] && [url.host isEqualToString:@"x-callback-url"] && [url.path hasPrefix:@"/vzero/auth/venmo/"];
-}
-
-+ (BOOL)isValidVenmoURL:(NSURL *)url andUnknownApplication:(NSString *)sourceApplication {
-    return sourceApplication == nil && [url.host isEqualToString:@"x-callback-url"] && [url.path hasPrefix:@"/vzero/auth/venmo/"];
 }
 
 @end