+++ UnitTests/BraintreeVenmoTests/BraintreeVenmoTests-Bridging-Header.h
@@ -1,1 +1,2 @@
 #import <BraintreeVenmo/BTVenmoDriver_Internal.h>
+#import <BraintreeVenmo/BTVenmoAppSwitchReturnURL.h>